requirements.txt
flask
flask-cors
mysql-connector-python
APScheduler>=3.10.0
bcrypt>=4.0.0
